added the ability to use session
