+++ helpers/getPasswordFrom.js
@@ -1,0 +1,21 @@
+const user = require("../database/models/user");
+
+function makeQuery(userID) {
+	return {
+		where: {
+			user_id: userID
+		}
+	};
+}
+
+module.exports = async userID => {
+	var password;
+	const query = makeQuery(userID);
+	await user
+		.findAll(query)
+		.catch(errorHandler)
+		.then(res => {
+			password = res[0].password;
+		});
+	return password;
+};
