+++ package.json
@@ -3,6 +3,7 @@
     "bcrypt": "^3.0.6",
     "curl": "^0.1.4",
     "express": "^4.16.4",
+    "express-session": "^1.16.2",
     "helmet": "^3.18.0",
     "htmlescape": "^1.1.1",
     "http-server": "^0.11.1",
